+++ academiclibrary/admin/sql/updates/mysql/1.sql
@@ -38,7 +38,7 @@ CREATE TABLE `#__al_trabalhos` (
 	`tra_titulo` VARCHAR(255) NOT NULL,
 	`tra_ano` INT(5) NOT NULL,
 	`tra_cat_id` INT(10) NOT NULL,
-	`tra_nota` DOUBLE(10,0) NOT NULL,
+	`tra_nota` DECIMAL(2,1) NOT NULL,
 	`tra_palavras_chaves` TEXT(65535) NOT NULL,
 	`tra_resumo` TEXT(65535) NOT NULL,
 	`tra_defesa_data` DATE NOT NULL,
@@ -91,6 +91,7 @@ INSERT INTO `#__al_docentes` (`doc_id`,`doc_nome`)
 VALUES (2, 'Macilon Araújo');
 INSERT INTO `#__al_docentes` (`doc_id`,`doc_nome`)
 VALUES (3, 'Catarina Costa');
+
 INSERT INTO `#__al_discentes` (`dis_id`, `dis_matricula`, `dis_nome`)
 VALUES (1, 20160300010, 'Mateus Costa');
